requirements.txt
Django==3.0.3
django-allauth==0.41.0
django-environ==0.4.5
psycopg2-binary==2.8.4
celery==4.4.0
